pyproject.toml
[tool.rye.scripts]
build = "docker build --tag=carderne/sigexport:dev ."
sig = "sigexport"

fmt = "rye fmt"
lint = "rye lint --fix"
check = "pyright"
test = "pytest"
all = { chain = ["fmt", "lint", "check", "test"] }

[project]
name = "signal-export"
description = "Export Signal conversations to Markdown and HTML"
readme = "README.md"
license = {text = "MIT License"}
requires-python = ">= 3.9"
keywords = ["backup", "chat", "export"]

authors = [
    {name = "Chris Arderne", email="chris@rdrn.me"},
]

dynamic = ["version"]

classifiers = [
  "Environment :: Console",
  "License :: OSI Approved :: MIT License",
  "Operating System :: Unix",
  "Programming Language :: Python",
]

dependencies = [
    "beautifulsoup4 ~= 4.11",
    "emoji ~= 2.0",
    "Markdown ~= 3.4",
    "typer[all] ~= 0.7",
    "pycryptodome ~= 3.20.0",
]

[tool.rye]
managed = true
generate-hashes = false
dev-dependencies = [
    "build",
    "pyright",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "ruff ~= 0.2",
    "setuptools-scm",
    "twine",
    "types-Markdown",
    "types-emoji",
]

[project.urls]
homepage = "https://github.com/carderne/signal-export"
repository = "https://github.com/carderne/signal-export"

[project.scripts]
sigexport = "sigexport.main:cli"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = ["src/"]

[tool.pdm.version]
source = "scm"

[tool.ruff]
target-version = "py39"
exclude = ["tests", "src/pysqlcipher3"]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "I",
    "U",
    "N",
    "E",
    "S",
    "T100",
    "A",
    "Q",
    "ANN",
]

[tool.ruff.lint.isort]
known-first-party = ["sigexport"]

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src/sigexport", "tests"]
ignore = ["src/pysqlcipher3"]
reportMissingImports = true
reportMissingParameterType = true
reportUnnecessaryTypeIgnoreComment = true
reportDeprecated = true
pythonVersion = "3.9"
